backend/requirements.txt
flask
flask-cors
praw
nltk
vaderSentiment
pandas
numpy
python-dotenv
